Rearrange file structure and initialize last message number for order stream
Continued rearranging file structure so bus related files are collected together.
The order message stream now gets initialized with the last message processed,
as read from the database as opposed to simply defaulting to the start of the stream.

+++ service-a/Bus/MessageStreamState.cs
@@ -4,7 +4,7 @@
     {
         public MessageStreamState(
             MessageStreamLocation streamLocation, 
-            int lastSuccessfullyProcessedMessage,
+            long lastSuccessfullyProcessedMessage,
             string description)
         {
             StreamLocation = streamLocation;
@@ -13,7 +13,7 @@
         }
 
         public MessageStreamLocation StreamLocation { get; }
-        public int LastSuccessfullyProcessedMessage { get; }
+        public long LastSuccessfullyProcessedMessage { get; }
         public string Description { get; }
     }
 }
